Hidden Units - Separate WS into subcomponent
