+++ addons/hidden_units/CfgFactionClasses.hpp
@@ -2,9 +2,6 @@ class CfgFactionClasses {
     class BLU_CTRG_F {
         displayName = "DEPRECATED";
     };
-    class BLU_UN_lxWS {
-        displayName = "DEPRECATED";
-    };
 
     // CUP Factions
     class CUP_B_USMC {
