+++ addons/hidden_units/CfgVehicles_Baseclasses.hpp
@@ -4,7 +4,6 @@
 class B_CTRG_Soldier_F;
 class B_CTRG_Soldier_3_F;
 class B_soldier_UAV_F;
-class B_ION_Soldier_SG_lxWS;
 
 // Vehicles
 class LSV_01_light_base_F;
@@ -15,94 +14,6 @@ class O_Truck_03_repair_F;
 class O_Truck_03_ammo_F;
 class O_Truck_03_fuel_F;
 class O_Truck_03_medical_F;
-
-// SFIA
-// Units
-class I_C_Soldier_Bandit_1_F;
-class I_C_Soldier_Bandit_2_F;
-class I_C_Soldier_Bandit_4_F;
-class I_C_Soldier_Bandit_5_F;
-class I_C_Soldier_Para_2_F;
-class I_crew_F;
-class I_G_medic_F;
-class I_G_Sharpshooter_F;
-class I_G_Soldier_exp_F;
-class I_officer_F;
-class I_pilot_F;
-class I_PMC_Soldier_01_lxWS;
-class I_SFIA_deserter_lxWS;
-class I_SFIA_soldier_unarmed_lxWS;
-class I_SFIA_survivor_lxWS;
-class I_Soldier_AA_F;
-class I_Soldier_AAA_F;
-class I_Soldier_AAT_F;
-class I_Soldier_AR_F;
-class I_Soldier_AT_F;
-class I_soldier_F;
-class I_Soldier_GL_F;
-class I_Soldier_repair_F;
-class I_Soldier_TL_F;
-
-// Vehicles
-class I_SFIA_APC_Tracked_02_30mm_lxWS;
-class I_SFIA_APC_Tracked_02_AA_lxWS;
-class I_SFIA_APC_Tracked_02_cannon_lxWS;
-class I_SFIA_APC_Wheeled_02_hmg_lxWS;
-class I_SFIA_APC_Wheeled_02_unarmed_lxWS;
-class I_SFIA_Heli_Attack_02_dynamicLoadout_lxWS;
-class I_SFIA_HMG_02_high_lxWS;
-class I_SFIA_HMG_02_lxWS;
-class I_SFIA_MBT_02_cannon_lxWS;
-class I_SFIA_Mortar_lxWS;
-class I_SFIA_Offroad_armed_lxWS;
-class I_SFIA_Offroad_AT_lxWS;
-class I_SFIA_Offroad_lxWS;
-class I_SFIA_Truck_02_aa_lxWS;
-class I_SFIA_Truck_02_Ammo_lxWS;
-class I_SFIA_Truck_02_box_lxWS;
-class I_SFIA_Truck_02_cargo_lxWS;
-class I_SFIA_Truck_02_covered_lxWS;
-class I_SFIA_Truck_02_flatbed_lxWS;
-class I_SFIA_Truck_02_fuel_lxWS;
-class I_SFIA_Truck_02_MRL_lxWS;
-class I_SFIA_Truck_02_transport_lxWS;
-class I_SFIA_ZU23_lxWS;
-class Offroad_01_armor_armed_lxWS;
-class Offroad_01_armor_AT_lxWS;
-class Offroad_01_armor_base_lxWS;
-class UAV_02_IED_Base_lxWS;
-
-// UNA
-class B_crew_F;
-class B_engineer_F;
-class B_Helipilot_F;
-class B_medic_F;
-class B_officer_F;
-class B_soldier_AR_F;
-class B_Soldier_F;
-class B_Soldier_GL_F;
-class B_soldier_LAT2_F;
-class B_Soldier_lite_F;
-class B_soldier_repair_F;
-class B_Soldier_TL_F;
-
-// Vehicles
-class APC_Wheeled_01_command_base_lxWS;
-class APC_Wheeled_02_hmg_base_lxws;
-class APC_Wheeled_02_unarmed_base_lxws;
-class B_Truck_01_mover_F;
-class B_Truck_01_ammo_F;
-class B_Truck_01_box_F;
-class B_Truck_01_fuel_F;
-class B_Truck_01_medical_F;
-class B_Truck_01_Repair_F;
-class B_Truck_01_transport_F;
-class B_Truck_01_covered_F;
-class B_MRAP_01_F;
-class Offroad_01_base_lxWS;
-class Offroad_01_armor_base_lxWS;
-class Heli_Transport_02_base_F;
-class I_MBT_03_cannon_F;
 
 // CUP
 // Units
