+++ addons/hidden_units/ws/CfgFactionClasses.hpp
@@ -1,0 +1,12 @@
+class CfgFactionClasses {
+    class BLU_UN_lxWS {
+        displayName = "DEPRECATED";
+    };
+    class IND_SFIA_lxWS {
+        displayName = "DEPRECATED";
+    };
+    class IND_TURA_lxWS {
+        displayName = "DEPRECATED";
+    };
+};
+
