+++ addons/hidden_units/ws/CfgGroups.hpp
@@ -1,0 +1,131 @@
+class CfgGroups {
+    class West {
+        // UNA; Replaced by UN
+        class BLU_UN_lxWS {
+            class Infantry {
+                name = "DEPRECATED";
+                delete BUNA_InfSquad_lxWS;
+                delete BUNA_InfSentry_lxWS;
+            };
+            class Motorized {
+                name = "DEPRECATED";
+                delete BUNA_MotInf_Team;
+                delete BUNA_MotInf_Reinforce;
+            };
+            class Mechanized {
+                name = "DEPRECATED";
+                delete BUNA_MechInfSquad;
+                delete BUNA_MechInf_Support;
+            };
+            class Armored {
+                name = "DEPRECATED";
+                delete BUNA_TankPlatoon;
+                delete BUNA_TankSection;
+            };
+        };
+        class BLU_TURA_lxWS {
+            class Infantry {
+                name = "DEPRECATED";
+                delete B_Tura_InfSquad_lxWS;
+                delete B_Tura_InfSentry_lxWS;
+                delete B_Tura_InfTeam_lxWS;
+            };
+            class Motorized {
+                name = "DEPRECATED";
+                delete B_Tura_MotInf_Team_lxWS;
+                delete B_Tura_MotInf_AT_lxWS;
+                delete B_Tura_MotInf_Reinforce_lxWS;
+            };
+        };
+    };
+    class East {
+        class OPF_SFIA_lxWS {
+            class Infantry {
+                name = "DEPRECATED";
+                delete OSFIA_InfSquad_lxWS;
+                delete OSFIA_InfSentry_lxWS;
+                delete OSFIA_InfTeam_lxWS;
+                delete OSFIA_HAF_InfTeam_AT_lxWS;
+                delete OSFIA_HAF_InfTeam_AA_lxWS;
+                delete OSFIA_InfSquad_Weapons_lxWS;
+            };
+            class Motorized {
+                name = "DEPRECATED";
+                delete OSFIA_MotInf_Team_lxWS;
+                delete OSFIA_MotInf_AT_lxWS;
+                delete OSFIA_MotInf_AA_lxWS;
+                delete OSFIA_MotInf_Reinforce_lxWS;
+            };
+            class Mechanized {
+                name = "DEPRECATED";
+                delete OSFIA_MechInf_AT;
+                delete OSFIA_MechInf_AA;
+            };
+            class Armored {
+                name = "DEPRECATED";
+                delete OSFIA_HAF_TankPlatoon_lxWS;
+                delete OSFIA_HAF_TankPlatoon_AA_lxWS;
+                delete OSFIA_HAF_TankSection_lxWS;
+            };
+        };
+        class OPF_TURA_lxWS {
+            class Infantry {
+                name = "DEPRECATED";
+                delete O_Tura_InfSquad_lxWS;
+                delete O_Tura_InfSentry_lxWS;
+                delete O_Tura_InfTeam_lxWS;
+            };
+            class Motorized {
+                name = "DEPRECATED";
+                delete O_Tura_MotInf_Team_lxWS;
+                delete O_Tura_MotInf_AT_lxWS;
+                delete O_Tura_MotInf_Reinforce_lxWS;
+            };
+        };
+    };
+    class Indep {
+        class IND_SFIA_lxWS {
+            class Infantry {
+                name = "DEPRECATED";
+                delete ISFIA_InfSquad_lxWS;
+                delete ISFIA_InfSentry_lxWS;
+                delete ISFIA_InfTeam_lxWS;
+                delete ISFIA_HAF_InfTeam_AT_lxWS;
+                delete ISFIA_HAF_InfTeam_AA_lxWS;
+                delete ISFIA_InfSquad_Weapons_lxWS;
+            };
+            class Motorized {
+                name = "DEPRECATED";
+                delete ISFIA_MotInf_Team_lxWS;
+                delete ISFIA_MotInf_AT_lxWS;
+                delete ISFIA_MotInf_AA_lxWS;
+                delete ISFIA_MotInf_Reinforce_lxWS;
+            };
+            class Mechanized {
+                name = "DEPRECATED";
+                delete ISFIA_MechInf_AT;
+                delete ISFIA_MechInf_AA;
+            };
+            class Armored {
+                name = "DEPRECATED";
+                delete ISFIA_HAF_TankPlatoon_lxWS;
+                delete ISFIA_HAF_TankPlatoon_AA_lxWS;
+                delete ISFIA_HAF_TankSection_lxWS;
+            };
+        };
+        class IND_TURA_lxWS {
+            class Infantry {
+                name = "DEPRECATED";
+                delete I_Tura_InfSquad_lxWS;
+                delete I_Tura_InfSentry_lxWS;
+                delete I_Tura_InfTeam_lxWS;
+            };
+            class Motorized {
+                name = "DEPRECATED";
+                delete I_Tura_MotInf_Team_lxWS;
+                delete I_Tura_MotInf_AT_lxWS;
+                delete I_Tura_MotInf_Reinforce_lxWS;
+            };
+        };
+    };
+};
