+++ addons/hidden_units/ws/CfgVehicles.hpp
@@ -1,0 +1,7 @@
+class CfgVehicles {
+    #include "baseclasses.hpp"
+
+    #include "ctrg.hpp"
+    #include "sfia.hpp"
+    #include "una.hpp"
+};
