+++ addons/hidden_units/ws/ctrg.hpp
@@ -1,0 +1,41 @@
+// CTRG
+class B_D_CTRG_Soldier_TL_lxWS: B_CTRG_Soldier_TL_tna_F {
+    scope = 1;
+    scopeCurator = 1;
+};
+class B_D_CTRG_soldier_UAV_lxWS: B_soldier_UAV_F {
+    scope = 1;
+    scopeCurator = 1;
+};
+class B_D_CTRG_Soldier_lxWS: B_CTRG_Soldier_tna_F {
+    scope = 1;
+    scopeCurator = 1;
+};
+class B_D_CTRG_Soldier_SG_lxWS: B_ION_Soldier_SG_lxWS {
+    scope = 1;
+    scopeCurator = 1;
+};
+class B_D_CTRG_Soldier_LAT2_lxWS: B_CTRG_Soldier_LAT2_tna_F {
+    scope = 1;
+    scopeCurator = 1;
+};
+class B_D_CTRG_Soldier_Medic_lxWS: B_CTRG_Soldier_Medic_tna_F {
+    scope = 1;
+    scopeCurator = 1;
+};
+class B_D_CTRG_Soldier_M_lxWS: B_CTRG_Soldier_M_tna_F {
+    scope = 1;
+    scopeCurator = 1;
+};
+class B_D_CTRG_Soldier_JTAC_lxWS: B_CTRG_Soldier_JTAC_tna_F {
+    scope = 1;
+    scopeCurator = 1;
+};
+class B_D_CTRG_Soldier_HG_lxWS: B_CTRG_Soldier_AR_tna_F {
+    scope = 1;
+    scopeCurator = 1;
+};
+class B_D_CTRG_Soldier_Exp_lxWS: B_CTRG_Soldier_Exp_tna_F {
+    scope = 1;
+    scopeCurator = 1;
+};
