+++ addons/hidden_units/ws/script_component.hpp
@@ -1,0 +1,4 @@
+#define COMPONENT hidden_units_ws
+#define COMPONENT_BEAUTIFIED Hidden Units WS
+#include "\x\tacgt\addons\main\script_mod.hpp"
+#include "\x\tacgt\addons\main\script_macros.hpp"
